+++ src/Hacks/Shortcuts/TogglePractice.cpp
@@ -1,5 +1,3 @@
-#pragma once
-
 #include <Geode/Geode.hpp>
 #include "../../Client/ButtonModule.hpp"